Move Paragraph to genhtml
